refactor/base structure and optimized to user entity
