+++ go.mod
@@ -2,4 +2,4 @@ module github.com/wizeline/CA-Microservices-Go
 
 go 1.22.0
 
-require github.com/lib/pq v1.10.9 // indirect
+require github.com/lib/pq v1.10.9
