--- internal/core/domain/entity/movie.go
@@ -1,18 +1,0 @@
-package entity
-
-import "time"
-
-type Movie struct {
-	ID        int
-	Name      string
-	Duration  time.Time
-	Release   time.Time
-	Genre     string
-	Directors []string
-	Actors    []string
-	Writers   []string
-	Country   string
-
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
