--- internal/core/infraestructure/handler/http.go
@@ -1,7 +1,0 @@
-package handler
-
-// HTTP controller for HTTP protocol.
-type HTTP interface {
-	// SetRoutes allocates routes into the given Group.
-	// SetRoutes(g *Group)
-}
