--- internal/core/infraestructure/handler/http_user_test.go
@@ -1,1 +1,0 @@
-package handler
