--- internal/core/infraestructure/repository/user_pg_test.go
@@ -1,1 +1,0 @@
-package repository
