+++ internal/domain/service/user.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/wizeline/CA-Microservices-Go/internal/core/domain/entity"
+	"github.com/wizeline/CA-Microservices-Go/internal/domain/entity"
 )
 
 type UserService interface {
