+++ internal/infraestructure/config/database.go
@@ -1,0 +1,1 @@
+package config
