+++ internal/infraestructure/middleware/jwt.go
@@ -1,0 +1,1 @@
+package middleware
