+++ internal/infraestructure/repository/error.go
