+++ internal/infraestructure/router/mux.go
@@ -1,0 +1,1 @@
+package router
