+++ internal/infraestructure/server/echo.go
@@ -1,0 +1,1 @@
+package server
